Add: Parse method to convert lines into ChatMessageObjects.
A static parse method created, this method reads a list of lines and builds ChatMessage objects.
This method also handles the validation of the data, checking of the format and also throws and InvalidMessageFormatException on errors related to invalid headings or unexpected end of lines or no empty lines between messages.